requirements.txt
astroid==2.8.4
black==21.9b0
click==8.0.3
Flask==2.0.2
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pylint==2.11.1
python-dotenv==0.19.1
regex==2021.10.23
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
Werkzeug==2.0.2
wrapt==1.13.2
